refactor pickers to use separate components
