+++ client/index.js
@@ -1,16 +1,15 @@
 import React from 'react'
 import { render } from 'react-dom'
-
 import { Provider } from 'react-redux'
 import { DevTools,
          DebugPanel,
          LogMonitor } from 'redux-devtools/lib/react'
 // import { createStore } from 'redux'
 
-import { Dashboard } from '../common/views'
-// import startSocketListener from './socketListener'
+import { App } from '../common/views'
 import anchorApp from '../common/state/reducers'
 // import * as actions from '../common/state/actions'
+// import startSocketListener from './socketListener'
 import configureStore from '../common/state/stores/configureStore'
 
 const initialState = window.__INITIAL_STATE__
@@ -19,7 +18,7 @@ const rootElement = document.getElementById('root')
 render(
   <div>
     <Provider store={store}>
-      <Dashboard />
+      <App />
     </Provider>
     <DebugPanel top right bottom>
       <DevTools store={store} monitor={LogMonitor} />
