+++ common/state/reducers/index.js
@@ -1,4 +1,5 @@
 import { combineReducers } from 'redux'
+
 // import user from './user'
 import dashboards from './dashboard/dashboards'
 import inputTypes from './inputTypes'
