+++ common/state/reducers/inputTypes.js
@@ -5,15 +5,19 @@ import * as types from '../constants/actionTypes'
 const initialState = {
   inputTypesById: {
     0: {
+      id: '0',
       title: 'search'
     },
     1: {
+      id: '1',
       title: 'list'
     },
     2: {
+      id: '2',
       title: 'resource'
     },
     3: {
+      id: '3',
       title: 'dashboard'
     }
   },
