+++ common/views/components/DashboardBody/ListPicker.js
@@ -1,4 +1,5 @@
 import React, { PropTypes } from 'react'
+
 import { List } from '../../'
 
 export const ListPicker = (
