+++ common/views/containers/App.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import { DashboardHeaderContainer,
+         DashboardBodyContainer } from '../'
+
+export const App = () => {
+  return (
+    <div>
+      <DashboardHeaderContainer />
+      <DashboardBodyContainer />
+    </div>
+  )
+}
