--- common/views/containers/Dashboard.js
@@ -1,22 +1,0 @@
-import React, { PropTypes } from 'react'
-import { DashboardHeaderContainer,
-         DashboardBodyContainer } from '../'
-
-export const Dashboard = () => {
-  return (
-    <div>
-      <DashboardHeaderContainer />
-      <DashboardBodyContainer />
-
-      {/*
-      <AddList
-        onAddListSubmit={ (dashboardId, title) => addList((dashboardId, title)) }
-      />
-      <ListContainer
-        lists={ dashboard.lists }
-        onAddSubmit={ (listId, url) => addResource(listId, url) }
-      /> */}
-    </div>
-  )
-}
-
