+++ common/views/containers/DashboardBodyContainer.js
@@ -1,5 +1,6 @@
 import React, { Component, PropTypes } from 'react'
 import { connect } from 'react-redux'
+
 import { DashboardBody } from '../'
 import { setCurrentList } from '../../state/actions'
