+++ common/views/index.js
@@ -1,16 +1,15 @@
-// Resources
-// =========
-export * from './components/AddResource'
-
-// Lists
-// =====
-export * from './components/AddList'
+// Containers
+// ==========
+export * from './containers/App'
+export { default as DashboardHeaderContainer } from './containers/DashboardHeaderContainer'
+export { default as DashboardBodyContainer } from './containers/DashboardBodyContainer'
 
 // Dashboard Header
 // ================
 export * from './components/DashboardHeader/DashboardHeader'
+export * from './components/DashboardHeader/Dashboard'
 export * from './components/DashboardHeader/DashboardPicker'
-export * from './components/DashboardHeader/InputTypes'
+export * from './components/DashboardHeader/InputType'
 export * from './components/DashboardHeader/InputTypesPicker'
 
 // Dashboard Body
@@ -18,11 +17,4 @@ export * from './components/DashboardHeader/InputTypesPicker'
 export * from './components/DashboardBody/DashboardBody'
 export * from './components/DashboardBody/List'
 export * from './components/DashboardBody/ListPicker'
-export * from './components/DashboardBody/ListContainer'
 export * from './components/DashboardBody/Resource'
-
-// Containers
-// ==========
-export * from './containers/Dashboard'
-export { default as DashboardHeaderContainer } from './containers/DashboardHeaderContainer'
-export { default as DashboardBodyContainer } from './containers/DashboardBodyContainer'
